+++ src/config/ApiEndPoints.ts
@@ -1,8 +1,8 @@
 export const API_BASE_URL = 'http://localhost:3001';
 
 export const API_ENDPOINTS = {
-  SIGNUP_EMAIL: `/auth/signup/email`,
-  SIGNIN_EMAIL: `/auth/signin/email`,
+  AUTH_SIGNUP_EMAIL: `/auth/signup/email`,
+  AUTH_SIGNIN_EMAIL: `/auth/signin/email`,
   HOME: `/home`,
   SHOP_ALL: `/shop/all`,
   SHOP: `/shop`,
