+++ src/types/shop.ts
@@ -1,4 +1,3 @@
-// 상품 인터페이스
 export interface Product {
   id: number;
   name: string;
@@ -8,13 +7,11 @@ export interface Product {
   soldCount: number;
 }
 
-// 카테고리 인터페이스
 export interface Category {
   id: number;
   name: string;
 }
 
-// 전체 데이터 인터페이스
 export interface ShopData {
   products: Product[];
   categories: Category[];
